Rename 'components/ConfirmedCasesByMunicipalitiesTable.vue' to 'components/index/CardsReference/ConfirmedCasesByMunicipalities/DataTable.vue'
